Remove iOS click trigger code
Why: I found that, on iOS browsers, when the user
touches an element and triggers a hover state, if
certain visual properties change as a result of a
CSS class toggle then the "click" event does not
occur on the element. The iOS click trigger code
acted on an assumption that is not true. It would
always trigger a click when hover was detected.